Rename package Mapper to mapper and DTO to dto
