+++ src/main/java/org/example/springboot25/dto/EventUpdateDTO.java
@@ -1,15 +1,12 @@
-package org.example.springboot25.DTO;
+package org.example.springboot25.dto;
 
 import java.time.OffsetDateTime;
 
-public class EventInputDTO {
-
+public class EventUpdateDTO {
     private String eventName;
     private String eventDescription;
     private String eventLocation;
     private OffsetDateTime eventDateTime;
-    private Long userId;
-
 
     public String getEventName() {
         return eventName;
@@ -41,13 +38,5 @@ public class EventInputDTO {
 
     public void setEventDateTime(OffsetDateTime eventDateTime) {
         this.eventDateTime = eventDateTime;
-    }
-
-    public Long getUserId() {
-        return userId;
-    }
-
-    public void setUserId(Long userId) {
-        this.userId = userId;
     }
 }
